fastapi-app/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
pypdf==3.1.0
python-jose==3.3.0
requests==2.28.0
